# backend/docs/swagger.yaml
basePath: /
definitions:
...
  models.CreateUserResponse:
    properties:
      message:
        type: string
    type: object
...
  models.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  models.ParentChildRequest:
    properties:
      child:
        $ref: '#/definitions/models.User'
      child_id:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      parent:
        allOf:
        - $ref: '#/definitions/models.User'
        description: Associations
      parent_id:
        type: integer
      status:
        description: pending, accepted, rejected
        type: string
      updated_at:
        type: string
    type: object
...
  models.User:
    properties:
      createdAt:
        type: string
      email:
        type: string
      id:
        type: integer
      passwordHash:
        type: string
      resetToken:
        description: Token de réinitialisation
        type: string
      resetTokenExpiry:
        description: Expiration du token de réinitialisation
        type: string
      role:
        type: string
      updatedAt:
        type: string
      username:
        type: string
    type: object
  models.UserResponse:
    properties:
      email:
        type: string
      id:
        type: integer
      role:
        type: string
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: contact@dratsisama.com
    name: Dr_AtsiSama
    url: http://dratsisama.com
  description: Documentation de l'API pour la gestion des kermesses scolaires avec
    fonctionnalités de stand, tombola, utilisateurs, etc.
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
  title: Kermisys API Documentation
  version: "1.0"
paths:
...
  /user:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Crée un nouvel utilisateur avec le nom d'utilisateur, l'adresse
        e-mail, et le rôle spécifiés
      parameters:
      - description: Nom d'utilisateur
        in: formData
        name: username
        required: true
        type: string
      - description: Adresse e-mail
        in: formData
        name: email
        required: true
        type: string
      - description: 'Rôle de l''utilisateur (ex: ''parent'', ''enfant'', ''organisateur'')'
        in: formData
        name: role
        required: true
        type: string
      - description: Mot de passe de l'utilisateur
        in: formData
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.CreateUserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      summary: Créer un nouvel utilisateur
      tags:
      - Utilisateurs
  /user/{username}:
    get:
      description: Récupère les informations d'un utilisateur à partir de son nom
        d'utilisateur
      parameters:
      - description: Nom d'utilisateur
        in: path
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      summary: Récupérer les détails d'un utilisateur
      tags:
      - Utilisateurs
  /user/children:
    get:
      description: Renvoie la liste de tous les enfants associés à un parent spécifique
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      summary: Liste des enfants associés à un parent
      tags:
      - Parent-Enfant
  /user/children/{child_id}:
    delete:
      description: Permet de désassocier un enfant d'un parent et de lui réassigner
        le rôle par défaut "parent"
      parameters:
      - description: ID de l'enfant
        in: path
        name: child_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      summary: Supprimer un enfant associé à un parent
      tags:
      - Parent-Enfant
  /user/parent-requests:
    post:
      consumes:
      - application/json
      description: Permet de créer une nouvelle demande pour lier un parent et un
        enfant en utilisant leurs noms d'utilisateur
      parameters:
      - description: Nom d'utilisateur du parent
        in: formData
        name: parent_username
        required: true
        type: string
      - description: Nom d'utilisateur de l'enfant
        in: formData
        name: child_username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.ParentChildRequest'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      summary: Ajouter une demande parent-enfant
      tags:
      - Parent-Enfant
  /user/parent-requests/accept/{id}:
    put:
      description: Permet d'accepter une demande existante pour lier un parent et
        un enfant
      parameters:
      - description: ID de la demande
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ParentChildRequest'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      summary: Accepter une demande parent-enfant
      tags:
      - Parent-Enfant
  /user/parent-requests/reject/{id}:
    put:
      description: Permet de refuser une demande existante pour lier un parent et
        un enfant
      parameters:
      - description: ID de la demande
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ParentChildRequest'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      summary: Refuser une demande parent-enfant
      tags:
      - Parent-Enfant
  /user/parent-requests/requests:
    get:
      description: Renvoie la liste de toutes les demandes en attente de traitement
        pour l'utilisateur connecté
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.ParentChildRequest'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      summary: Liste des demandes parent-enfant pour l'utilisateur connecté
      tags:
      - Parent-Enfant
...
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"